# backend/src/docs/api/openapi.yaml
openapi: 3.1.1
info:
  title: Commons Fabric API
  version: 0.1.0
  description: |-
    Events published by participating organizations.

    Errors use RFC 9457 problem details, served as `application/problem+json`.

    Two error responses apply to every path but are deliberately not attached to
    any operation, because they are returned precisely when a request matches no
    documented operation - so there is no operation under which to list them:

    - `404` - the requested path does not appear in this document.
    - `405` - the path is documented, but not for the method used.
  license:
    name: GNU Affero General Public License v3.0 or later
    identifier: AGPL-3.0-or-later
servers:
  - url: http://localhost:3000/v1 #TODO: use environment variables for this? - requires some setup in the code that parses this file. [Where?]
security: [] #TODO: figure out how to configure this from @marcus-wrrn once auth is implemented
...
paths:
  /auth/create-user:
    post:
      security: [] #public - self-registration -> might require some form of security later, keeping public for now 
      operationId: createUser
      tags: [Auth]
      summary: Create a user account
      description: >-
        Registers a new user under the given organization. `email` must be
        unique - a request reusing an existing address is rejected with 409.
        `password` is received as plaintext over TLS and hashed server-side
        (argon2id) before storage; it is never echoed back in a response. On
        success a signed, httpOnly session cookie is set, logging the new user
        in immediately.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Created
          headers:
            Location:
              description: URI of the newly created user.
              required: true
              schema: { type: string, format: uri-reference }
            Set-Cookie:
              description: Signed, httpOnly session cookie (see src/utils/userSessions.ts).
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/Conflict' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        default: { $ref: '#/components/responses/Problem' }
...
components:
...
  schemas:
    Timestamp:
      type: string
      format: date-time
    OrganizationId:
      type: integer
      minimum: 1
      description: Identifier of an organization (an `organizations` row).
    UserId:
      type: integer
      minimum: 1
      description: Identifier of a user.
...
    UserCore:
      type: object
      description: The client-supplied half of a user. Mirrors the `users` table.
      required: [fullname, email, organizationId]
      properties:
        fullname:
          type: string
          minLength: 1
          pattern: '\S'
          description: >-
            The `pattern` matches the `length(trim(fullname)) > 0` CHECK in
            schema.dbml.
        email:
          type: string
          format: email
          maxLength: 255
          pattern: '^[^@\s]+@[^@\s]+\.[^@\s]+$'
          description: >-
            Must be unique across users; a request reusing an existing address
            is rejected with 409. `pattern` mirrors the DBML CHECK and `format`
            applies ajv's own rule - see the note on `Organization.contact`.
        organizationId: { $ref: '#/components/schemas/OrganizationId' }
    UserCreate:
      allOf:
        - $ref: '#/components/schemas/UserCore'
        - type: object
          required: [password]
          properties:
            password:
              type: string
              minLength: 8
              writeOnly: true
              description: >-
                Plaintext password, sent once over TLS and hashed (argon2id)
                server-side before storage. Mirrors the
                `length(password) > 0` CHECK in schema.dbml, with a minimum
                length enforced at the application layer.
      description: >-
        Unknown properties are rejected rather than ignored, so a request that
        tries to smuggle in `id` or `createdAt` (or any other server-assigned
        field) fails validation instead of being silently dropped.
      unevaluatedProperties: false
    User:
      allOf:
        - $ref: '#/components/schemas/UserCore'
        - type: object
          required: [id, createdAt]
          properties:
            id: { $ref: '#/components/schemas/UserId', readOnly: true }
            createdAt: { $ref: '#/components/schemas/Timestamp', readOnly: true }
...
    Problem:
      type: object
      description: RFC 9457 problem details.
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, default: 'about:blank' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
    ValidationProblem:
      description: >-
        A `Problem` carrying the RFC 9457 `errors` extension. `errors` is present
        whenever the failure can be attributed to specific fields, and absent
        when it cannot - a body that is not parseable as JSON, for instance, has
        no field to point at.
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            errors:
              type: array
              minItems: 1
              items:
                type: object
                required: [pointer, detail]
                properties:
                  pointer:
                    type: string
                    description: >-
                      JSON Pointer to the offending value, prefixed with the part
                      of the request it came from - e.g. `/body/startsAt`,
                      `/query/startDate`, `/params/eventId`.
                  detail: { type: string }
  responses:
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationError:
      description: >-
        The request failed validation - body, query, or path parameters.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ValidationProblem' }
    UnsupportedMediaType:
      description: The request body was sent as something other than `application/json`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: >-
        The request conflicts with the current state of the resource - e.g. an
        `email` that is already registered to another user.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...